requirements.txt
streamlit>=1.40.0
gunicorn>=25.1.0
supabase>=2.0.0
google-generativeai>=0.8.0
nanoid>=2.0.0
pandas>=2.0.0
plotly>=5.18.0
Pillow>=10.0.0
qrcode>=7.4.0
python-dotenv>=1.0.0
